create MusicLyricForm.cs
