+++ MusicLyricApp/BlurForm.cs
@@ -1,13 +1,23 @@
-﻿using System.Windows.Forms;
-using MusicLyricApp.Bean;
+﻿using MusicLyricApp.Bean;
 
 namespace MusicLyricApp
 {
-    public partial class BlurForm : Form
+    public partial class BlurForm : MusicLyricForm
     {
+        private SearchResultVo _searchResultVo;
+        
         public BlurForm(SearchResultVo searchResultVo)
         {
+            _searchResultVo = searchResultVo;
+            
             InitializeComponent();
+            
+            AfterInitializeComponent();
+        }
+
+        private void AfterInitializeComponent()
+        {
+            
         }
     }
 }
