+++ MusicLyricApp/MusicLyricApp.csproj
@@ -269,6 +269,7 @@
     <Compile Include="MainForm.Designer.cs">
       <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
+    <Compile Include="MusicLyricForm.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SettingForm.cs">
@@ -299,6 +300,9 @@
     <Compile Include="Utils\RomajiUtils.cs" />
     <Compile Include="Utils\SrtUtils.cs" />
     <Compile Include="Utils\XmlUtils.cs" />
+    <EmbeddedResource Include="BlurForm.resx">
+      <DependentUpon>BlurForm.cs</DependentUpon>
+    </EmbeddedResource>
     <EmbeddedResource Include="MainForm.resx">
       <DependentUpon>MainForm.cs</DependentUpon>
     </EmbeddedResource>
